breaking down of Get function and seperating user interaction from services now ready for testing(hope so :)

+++ backend/models/order.go
@@ -7,13 +7,13 @@ import (
 type Order struct {
 	OrderId    int
 	Products   []OrderItem
-	TotalPrice float32
+	TotalPrice float64
 	IsCredit   bool
 }
 
 type OrderItem struct { //to track only the necessary changes and keep the details current
 	ProductId int
-	Price     float32
+	Price     float64
 	Quantity  int
 	P         *Product
 }
@@ -36,26 +36,26 @@ func (o *Order) AddProduct(p *Product, Quantity int) bool {
 	for i, item := range o.Products { //checks whether same Product is already in order and so adds it in existing order details
 		if item.ProductId == p.ID {
 			o.Products[i].Quantity += Quantity
-			o.TotalPrice += p.Price * float32((Quantity))
+			o.TotalPrice += p.Price * float64((Quantity))
 			return true
 		}
 	}
 	newProduct := OrderItem{ // to add new Product in order list as no already existing Product found in order
 		ProductId: p.ID,
-		Price:     float32(p.Price),
+		Price:     float64(p.Price),
 		Quantity:  Quantity,
 		P:         p,
 	}
 	o.Products = append(o.Products, newProduct)
-	o.TotalPrice += p.Price * float32((Quantity))
+	o.TotalPrice += p.Price * float64((Quantity))
 	return true
 }
 
-func (o *Order) CalculateTotal() float32 {
+func (o *Order) CalculateTotal() float64 {
 
 	o.TotalPrice = 0.0
 	for _, item := range o.Products {
-		o.TotalPrice += float32(item.Price) * float32(item.Quantity)
+		o.TotalPrice += float64(item.Price) * float64(item.Quantity)
 	}
 	return o.TotalPrice
 }
